YTDB-523: Extract shared versioned put/remove into VersionedIndexOps
BTreeSingleValueIndexEngine and BTreeMultiValueIndexEngine had ~80 lines
of structurally identical versioning logic (put with snapshot pairs,
tombstone removal, count delta accumulation). Extract into static helpers
in VersionedIndexOps to keep both engines synchronized during future
changes.
